+++ src/js/mixin/internal/slideshow-animations.js
@@ -20,7 +20,7 @@ export default {
 };
 
 export function translated(el) {
-    return Math.abs(css(el, 'transform').split(',')[4] / el.offsetWidth) || 0;
+    return Math.abs(css(el, 'transform').split(',')[4] / el.offsetWidth);
 }
 
 export function translate(value = 0, unit = '%') {